+++ lib/roken/base64-test.c
@@ -33,7 +33,7 @@
 
 #include <config.h>
 
-#include "roken.h"
+#include <roken.h>
 #include <base64.h>
 
 int
@@ -57,7 +57,7 @@ main(int argc, char **argv)
     };
     for(t = tests; t->data; t++) {
 	char *str;
-	int len;
+	size_t len;
 	len = rk_base64_encode(t->data, t->len, &str);
 	if(strcmp(str, t->result) != 0) {
 	    fprintf(stderr, "failed test %d: %s != %s\n", numtest,
